# gh-138407: avoid useless join operation when calculate the hash vaule for pathlib.Path