--- a/Misc/NEWS.d/next/Library/2025-09-08-19-04-23.gh-issue-138407.BqNAfi.rst
+++ b/Misc/NEWS.d/next/Library/2025-09-08-19-04-23.gh-issue-138407.BqNAfi.rst
@@ -1,0 +1,1 @@
+Avoid useless join operation when calculate the hash vaule for pathlib.Path.