Signed-off-by: Manjusaka <me@manjusaka.me><!--

Try to fix #138407 

The benchmark:

Before this patch

```text
╰─ python -m pyperf timeit -s 'from pathlib import Path' 'p = hash(Path("/tmp/foo/bar/12345"))'
.....................
WARNING: the benchmark result may be unstable
* Not enough samples to get a stable result (95% certainly of less than 1% variation)

Try to rerun the benchmark with more runs, values and/or loops.
Run 'python -m pyperf system tune' command to reduce the system jitter.
Use pyperf stats, pyperf dump and pyperf hist to analyze results.
Use --quiet option to hide these warnings.

Mean +- std dev: 3.76 us +- 0.11 us
```

After

```text
╰─ python -m pyperf timeit -s 'from pathlib import Path' 'p = hash(Path("/tmp/foo/bar/12345"))'
.....................
Mean +- std dev: 2.47 us +- 0.04 us
```