@barneygale 2025-09-08
To add another wrinkle: sometimes paths are instantiated with their normalized string representation already set, e.g. this is true for the results of `Path.iterdir()`. I suspect your patch will be slower for something like `set(big_dir.iterdir())` (but I might be wrong).