@Zheaoli 2025-09-13
I have made a full benchmark test

The following is the result

## 🧪 Available Tests

- **Short Single-Level Path Benchmark**: Tests hash performance on simple paths like `"/abc"`
  - `bench_single_level_path.py` - Original pathlib implementation
  - `bench_single_level_path_new.py` - Optimized pathlib implementation

- **Long Single-Level Path Benchmark**: Tests hash performance on 512-character paths
  - `bench_long_single_level_path.py` - Original pathlib implementation
  - `bench_long_single_level_path_new.py` - Optimized pathlib implementation

- **Deep Path Benchmark**: Tests hash performance on 512-level deep paths (`/a/b/c/...`)
  - `bench_deep_path.py` - Original pathlib implementation
  - `bench_deep_path_new.py` - Optimized pathlib implementation

- **Deep Long Path Benchmark**: Tests hash performance on 512-level deep paths with 10-character names (`/abcdefghij/klmnopqrst/...`)
  - `bench_deep_long_path.py` - Original pathlib implementation
  - `bench_deep_long_path_new.py` - Optimized pathlib implementation

## =� Benchmark Details

The benchmark tests hash performance on single-level paths (`"/abc"`) by:
- Creating 100,000 Path objects in a loop
- Computing hash for each Path object
- Using `pyperf` for statistical accuracy

## =� Performance Results

### Local Benchmark Results

```
+-----------+-------------------+-----------------------+
| Benchmark | single_level_path | single_level_path_new |
+===========+===================+=======================+
| bench     | 197 ms            | 137 ms: 1.44x faster  |
+-----------+-------------------+-----------------------+
```

**Summary**: The optimized implementation shows a **1.44x performance improvement** (44% faster) for short single-level path hashing operations.

### Long Path Benchmark Results

```
+-----------+------------------------+----------------------------+
| Benchmark | long_single_level_path | long_single_level_path_new |
+===========+========================+============================+
| bench     | 241 ms                 | 180 ms: 1.34x faster       |
+-----------+------------------------+----------------------------+
```

**Summary**: The optimized implementation shows a **1.34x performance improvement** (34% faster) for long single-level path hashing operations.

### Deep Path Benchmark Results

```
+-----------+-----------+------------------------+
| Benchmark | deep_path | deep_path_new          |
+===========+===========+========================+
| bench     | 1.19 sec  | 1.07 sec: 1.12x faster |
+-----------+-----------+------------------------+
```

**Summary**: The optimized implementation shows a **1.12x performance improvement** (12% faster) for 512-level deep path hashing operations.

### Deep Long Path Benchmark Results

```
+-----------+----------------+------------------------+
| Benchmark | deep_long_path | deep_long_path_new     |
+===========+================+========================+
| bench     | 1.90 sec       | 2.02 sec: 1.06x slower |
+-----------+----------------+------------------------+
```

**Summary**: The optimized implementation shows a **1.06x performance regression** (6% slower) for 512-level deep paths with 10-character component names.

### Operation Pattern Benchmark Results

Testing different operation patterns on short single-level paths (`/abc`):

#### 1. hash(p) only - **1.44x faster** ✅
```
+-----------+-------------------+-----------------------+
| Benchmark | single_level_path | single_level_path_new |
+===========+===================+=======================+
| bench     | 197 ms            | 137 ms: 1.44x faster  |
+-----------+-------------------+-----------------------+
```

#### 2. str(p) only - **Equal performance** ✅
```
Benchmark hidden because not significant (1): bench
```

#### 3. str(p); hash(p) - **1.12x faster** ❌ (Expected: Slower)
```
+-----------+----------+----------------------+
| Benchmark | str_hash | str_hash_new         |
+===========+==========+======================+
| bench     | 213 ms   | 190 ms: 1.12x faster |
+-----------+----------+----------------------+
```

#### 4. hash(p); str(p) - **1.06x faster** ❌ (Expected: Slower)
```
+-----------+----------+----------------------+
| Benchmark | hash_str | hash_str_new         |
+===========+==========+======================+
| bench     | 201 ms   | 189 ms: 1.06x faster |
+-----------+----------+----------------------+
```