@barneygale 2025-09-18
I'm struggling to understand how the new code can be faster for `str(p); hash(p)` - the first part is unchanged, and the second part was effectively `hash(known_string.lower())` in the old code, whereas the new code seems to do much more. Is it the attribute lookup miss on `_str_normcase`?