@Zheaoli 2025-09-22
> Would you mind running that through your benchmarks please?

Sorry for reply late. I'm working on PyCon China 2025 last week. I will update this PR ASAP I can